Added a few more TODO items to get done by Jerry, Travis, or me.
but the real update:  is that a pointer to the "User" object is passed to CoreMechanics (which because this project is due monday and i have TONS of other projects that also are due, is now acting as the main engine for the game because I'm too lazy to actually move all of its game contents to its own class IM SORRY, i will move it if i plan to expand this game after school is finished lol). BECAUSE the user object pointer is passed to core mechanics, coremecahnics can now update the users inventory and pull information about the player of the game, this is vital for gameplay and since core mechanics will be our engine cause im lazy, its a necessary adjustment.

other update: The "animationLoading" function in core mechanics is overloaded so that you can also pass text to be displayed during loading time if you want.

also some other small things that no one cares about.

will anyone even read these commits? who even knows lol